fix admin delete record double button
<!-- ps-id: dac607a8-3ea8-45ef-8718-8ffa3a9d0bdb -->
